Feat: 포스트 뷰어 생성 및 더미 데이터 변경
#62
