+++ src/components/CKEditor5/Editor.css
@@ -17,6 +17,12 @@ ol {
   @apply min-h-[320px] max-h-[600px];
 }
 
+.ck-content.ck-read-only {
+  @apply px-0;
+  @apply min-h-0;
+  @apply border-none;
+}
+
 .ck.ck-editor__main > .ck-editor__editable.ck-focused {
   @apply border-gray-400 shadow-none;
 }
